requirements.txt
SQLAlchemy~=2.0.36
python-dotenv~=1.0.1
fastapi~=0.115.6
python-jose~=3.3.0
pydantic~=2.10.3